test_pr/pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "test_pr"
version = "0.1.0"
description = "Custom authentication & RBAC authorization on DRF without django.contrib.auth"
requires-python = ">=3.11"
dependencies = [
  "Django>=5.0",
  "djangorestframework>=3.15",
  "PyJWT>=2.9",
  "passlib[bcrypt]>=1.7.4",
  "python-dotenv>=1.0",
  "psycopg[binary]>=3.2",
  "dj-database-url>=2.2"
]

[tool.setuptools.packages.find]
include = ["test_pr", "accounts", "access", "resources"]